+++ .gitignore
@@ -59,3 +59,8 @@ config/environments/*.local.yml
 /node_modules
 yarn-debug.log*
 .yarn-integrity
+
+public/uploads/*
+
+# Ignore application configuration
+/config/application.yml
