+++ app/assets/stylesheets/custom.scss
@@ -40,3 +40,7 @@
     color: $state-danger-text;
   }
 }
+.picture-answer{
+  height: 100px;
+  width: 120px;
+}
