+++ app/uploaders/picture_uploader.rb
@@ -1,0 +1,6 @@
+class PictureUploader < CarrierWave::Uploader::Base
+  storage :file
+  def store_dir
+    "uploads/#{model.class.to_s.underscore}/#{mounted_as}/#{model.id}"
+  end
+end
