+++ app/views/admin/listening/conversation_questions/_question_type.html.erb
