+++ app/views/admin/listening/picture_questions/_question_type.html.erb
@@ -1,0 +1,38 @@
+<tr>
+  <td class="center">
+    <label class="pos-rel">
+      <input type="checkbox" class="ace" />
+      <span class="lbl"></span>
+    </label>
+  </td>
+  <td class="center">
+    <div class="action-buttons">
+      <a href="#" class="green bigger-140 show-details-btn" title="Show Details">
+        <i class="ace-icon fa fa-angle-double-down"></i>
+        <span class="sr-only"><%= t ".detail" %></span>
+      </a>
+    </div>
+  </td>
+  <td>
+    <audio class="html5_audio" controls=""><source src="<%=question_type.content%>" type="audio/mpeg">
+      <%= t ".browser" %>
+    </audio>
+  </td>
+  <td>
+    <div class="hidden-sm hidden-xs btn-group">
+      <a href="<%=edit_admin_listening_picture_question_path question_type%>" role="button" data-toggle="modal">
+        <button class="btn btn-xs btn-info">
+          <i class="ace-icon fa fa-pencil bigger-120"></i>
+        </button>
+      </a>
+      <% if not_belong_to_exam(question_type) %>
+        <%= link_to admin_listening_listening_question_path(question_type), method: :delete, data: { confirm: t(".confirm") } do %>
+          <button class="btn btn-xs btn-danger">
+            <i class="ace-icon fa fa-trash-o bigger-120">
+            </i>
+          </button>
+        <% end %>
+      <% end %>
+    </div>
+  </td>
+</tr>
