+++ config/initializers/assets.rb
@@ -27,3 +27,5 @@ Rails.application.config.assets.precompile += %w( user/course_responsive.css )
 Rails.application.config.assets.precompile += %w( user/plugins/parallax-js-master/parallax.min.js )
 Rails.application.config.assets.precompile += %w( user/plugins/progressbar/progressbar.min.js )
 Rails.application.config.assets.precompile += %w( user/course.js )
+Rails.application.config.assets.precompile += %w( picture_upload.js )
+Rails.application.config.assets.precompile += %w( table_detail.js )
