+++ config/locales/en.yml
@@ -56,19 +56,19 @@ en:
           add: "Add"
           detail: "Details"
           file: "FILE Nghe"
+        destroy:
+          flash: "Question deleted"
+      conversation_questions:
         question_type:
           detail: "Details"
           browser: "Your browser does not support the audio element."
           answer: "answer:"
           delete: "Delete"
           confirm: "You sure?"
-        destroy:
-          flash: "Question deleted"
-      conversation_questions:
         question:
           answer: "answer:"
         new:
-          title: "Create a new rewrite question"
+          title: "Create a new conversation question"
           question: "Question"
           answer: "Answer"
           close: "Close"
@@ -78,7 +78,7 @@ en:
           is_answer: "Is answer"
           audio_message: "Your browser not support"
         edit:
-          title: "Edit a rewrite question"
+          title: "Edit a conversation question"
           question: "Question"
           answer: "Answer"
           close: "Close"
@@ -92,8 +92,38 @@ en:
         update:
           flash_success: "Update success"
       picture_questions:
+        question_type:
+          detail: "Details"
+          browser: "Your browser does not support the audio element."
+          answer: "answer:"
+          delete: "Delete"
+          confirm: "You sure?"
         question:
           answer: "answer:"
+        new:
+          title: "Create a new picture question"
+          question: "Question"
+          answer: "Answer"
+          close: "Close"
+          submit: "Submit"
+          file: "Link file"
+          option: "Option"
+          is_answer: "Is answer"
+          audio_message: "Your browser not support"
+        edit:
+          title: "Edit a picture question"
+          question: "Question"
+          answer: "Answer"
+          close: "Close"
+          submit: "Submit"
+          file: "Link file"
+          option: "Option"
+          is_answer: "Is answer"
+          audio_message: "Your browser not support"
+        create:
+          flash_success: "Create success"
+        update:
+          flash_success: "Update success"
       listening_exams:
         index:
           part1: "Part 1: Picture listening"
