+++ db/migrate/20180817204827_add_picture_to_answer.rb
@@ -1,0 +1,5 @@
+class AddPictureToAnswer < ActiveRecord::Migration[5.2]
+  def change
+    add_column :answers, :picture, :string
+  end
+end
